% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RAData.R
\docType{package}
\name{RAData}
\alias{RAData}
\alias{RAData-package}
\title{RAData}
\description{
a description of the package.
}
\details{
RAData
A data package for study RAData

Additional details.
}
\seealso{
\link{mydataset}
}

